Move smoke tests into poll test module
